Added connection vaidation logic

+++ Public/Remove-PSPersonalVault.ps1
@@ -1,12 +1,14 @@
 function Remove-PSPersonalVault {
     [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'High')]
-    param (
+    param (        
         [switch] $Force
     )
     
     process {
-        if ($Force.IsPresent -or $PSCmdlet.ShouldProcess("Peronal Vault", "Remove-PSPersonalVault")) {
-            _clearPersonalVault
-        }
+        if (_isValidConnection (_getConnectionObject)) {
+            if ($Force.IsPresent -or $PSCmdlet.ShouldProcess("Peronal Vault", "Remove-PSPersonalVault")) {
+                _clearPersonalVault
+            }
+        } else { _connectionWarning }
     }
 }
