+++ Public/Remove-PSSecret.ps1
@@ -1,11 +1,7 @@
 function Remove-PSSecret {
     [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'High')]
     param (
-        [Parameter(
-            Mandatory = $true,
-            Position = 0,
-            ValueFromPipeline = $true,
-            ValueFromPipelineByPropertyName = $true)]
+        [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
         [ArgumentCompleter([NameCompleter])]
         [ValidateNotNullOrEmpty()]
         [string] $Name,
@@ -14,13 +10,14 @@ function Remove-PSSecret {
     )
     
     process {
-        if (!(_isNameExists $Name)) { Write-Warning "Couldn't find the value for given Name '$Name'; Pass the correct value and try again." }
+        if (_isValidConnection (_getConnectionObject)) {
+            if (!(_isNameExists $Name)) { Write-Warning "Couldn't find the value for given Name '$Name'; Pass the correct value and try again." }
 
-        else {
-            if ($Force -or $PSCmdlet.ShouldProcess($Name, "Remove-Secret")) {
-
-                Invoke-SqliteQuery -DataSource (_getDbPath) -Query "DELETE FROM _ WHERE Name = '$Name'"                
+            else {
+                if ($Force -or $PSCmdlet.ShouldProcess($Name, "Remove-Secret")) {
+                    Invoke-SqliteQuery -DataSource (_getDbPath) -Query "DELETE FROM _ WHERE Name = '$Name'"
+                }
             }
-        }
+        } else { _connectionWarning }
     }
 }
